tr2/objects: split cultist module
