+++ src/tr2/game/objects/creatures/cultist_1.h
@@ -3,5 +3,3 @@
 void Cultist1_Setup(void);
 void Cultist1A_Setup(void);
 void Cultist1B_Setup(void);
-void Cultist2_Setup(void);
-void Cultist3_Setup(void);
