+++ src/tr2/game/objects/creatures/cultist_2.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Cultist2_Setup(void);
