+++ src/tr2/game/objects/creatures/cultist_3.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Cultist3_Setup(void);
