+++ src/tr2/game/objects/creatures/cultist_common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "global/const.h"
+
+#define CULTIST_RADIUS (WALL_L / 10) // = 102
