+++ src/tr2/game/objects/setup.c
@@ -10,7 +10,9 @@
 #include "game/objects/creatures/big_spider.h"
 #include "game/objects/creatures/bird.h"
 #include "game/objects/creatures/bird_guardian.h"
-#include "game/objects/creatures/cultist.h"
+#include "game/objects/creatures/cultist_1.h"
+#include "game/objects/creatures/cultist_2.h"
+#include "game/objects/creatures/cultist_3.h"
 #include "game/objects/creatures/diver.h"
 #include "game/objects/creatures/dog.h"
 #include "game/objects/creatures/dragon.h"
