+++ src/tr2/meson.build
@@ -140,7 +140,9 @@ dll_sources = [
   'game/objects/creatures/big_spider.c',
   'game/objects/creatures/bird.c',
   'game/objects/creatures/bird_guardian.c',
-  'game/objects/creatures/cultist.c',
+  'game/objects/creatures/cultist_1.c',
+  'game/objects/creatures/cultist_2.c',
+  'game/objects/creatures/cultist_3.c',
   'game/objects/creatures/diver.c',
   'game/objects/creatures/dog.c',
   'game/objects/creatures/dragon.c',
